requirements.txt
asgiref==3.8.1
autopep8==2.3.2
Django==5.1.5
django-cors-headers==4.7.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-tracking==1.5.0
gprof2dot==2024.6.6
pillow==11.1.0
pycodestyle==2.12.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
